ShopNow/requirements.txt
Flask>=2.1.0
Flask-WTF>=1.0.0
Flask-Login>=0.6.0
Flask-Migrate>=3.1.0
Flask-SQLAlchemy>=3.0.0
PyMySQL>=1.0.2
python-dotenv>=0.19.0
Werkzeug>=2.0.0
Pillow>=9.0.0
pytest>=7.0.0
email-validator>=1.3.0

